requirements.txt
absl-py==2.1.0
altair==4.2.0
beautifulsoup4==4.12.2
click==8.1.3
cloudpickle==3.1.1
fastjsonschema==2.16.3
google-auth==2.37.0
google-generativeai==0.8.4
gym==0.26.2
h5py==3.12.1
httpx==0.24.0
huggingface-hub==0.28.1
ipykernel==6.22.0
ipython==8.12.3
jupyter==1.0.0
keras==2.10.0
matplotlib==3.9.1
numpy==1.26.4
pandas==2.2.3
pillow==9.5.0
scikit-learn==1.1.2
scipy==1.10.0
tensorflow==2.10.0
streamlit
